[#13] Add warnings about puppetlabs_spec_helper
As described in the referenced issue, puppetlabs_spec_helper had it's own
:syntax rake task which conflicted with ours. As of puppetlabs_spec_helper
0.8.0 (just released) it now depends on this project instead.

The additional `syntax:check_puppetlabs_spec_helper` task added to the
prereqs for :syntax will raise a STDERR warning if we find that
puppetlabs_spec_helper <= 0.7.0 has set its own actions for this task.

By being a prereq task it will get executed at runtime and correctly detect
it regardless of whether puppetlabs_spec_helper was loaded before or after
us. The task has no `desc` in order to prevent it from being listed and
cluttering the output of `rake -T`.

I've also added an entry to README for users that won't necessarily be using
the latest version of puppet-syntax.