Add check for non-ASCII characters
Manifests and templates that contain non-ASCII characters have been known to
break Puppet with the mysterious error message:

    invalid byte sequence in US-ASCII

We have several examples of this in our (currently internal) repo:

- gds/puppet@63b36f93bf75a848e2125008aa1e880c5861cf46
- gds/puppet@6ce832168a31d8794eafa23d09c7f2d8bc59e46b
- gds/puppet@9c5943f1d9e4f2c69929db67e631f833c6730d2f

Add a new task to prevent this from happening in future. This is an
iteration on @jennyd's original awk implementation in gds/puppet#2255.