Let Array#<< modify default configuration values
Previously, config reader methods would return a new array for the default
value of paths, so if a user did:

  PuppetSyntax.exclude_paths << "spec/fixtures/**/*.pp"

e.g. in puppetlabs_spec_helper, then the new path would get appended to the
new array and not change the actual configuration.  Now the array is set as
the default first, so in-place modifications of arrays are preserved.