Add message about Puppet 4 and future parser
The future parser setting no longer has an effect on Puppet 4.0.0 as there
is no longer a future parser. It is now just known as the "parser" in
general terms. While the `parser = future` setting may still be set, it
will be ignored by the current releases of Puppet 4.

This commit adds a warning message about setting the `future_parser`
setting when Puppet 4 is present. This message is simply to inform the
user that the option had been set but is no longer effective.

For the time being there is no future parser in Puppet 4. In the event
that a future parser is reintroduced, this code should be removed. For
now, we're still a long way off from Puppet 5. Overall the chances of another
DSL change the extent to which we saw with Puppet 4 is very unlikely
(but not impossible of course).