Add support for EPP templates
Puppet 4 now supports a new template language, the EPP (embedded puppet)
template.

This commit adds support for EPP templates. It does so by extending the
`PuppetSyntax::Templates` class. To validate the templates, the Puppet
EPP parser itself is being directly called in a manner very similar to
how the native `puppet epp validate` command works.

Since the EPP templates my be stored along side ERB templates in the
same location in the module, this raised a bit of an issue. Technically
it is not necessary to add the '.erb' or '.epp' extensions to the files,
though it is generally good practice to do so. This new code will look
for an explicit '.epp' extension and assume that anything else is meant
to be processes as an ERB template. There is one exception to this way
of handling templates. A new option, the `epp_only` options, has been
added. This option will cause all templates to be treated as EPP. This
is useful in the event a module wants to fully adopt the Puppet 4 way
only and not allow ERB.