Update Puppet 4 related test
The spec tests contain some tests that should not run unless the testing
suite is run against Puppet 4. This is largely being handled with
conditionals (though thats sort of ugly). The current unit tests for the
`PuppetSyntax::Manifests` class is missing a conditional around a test
meant only to run in Puppet 4.

This commit adds the conditional and the tests now complete
successfully.