Ignore future_parser in Puppet 4
This commit follows up on b7c2b1c in making it so that the future parser
setting is ignored when this project is running against Puppet 4.
Originally only an info message was added, but it appears that without
the setting being known to Puppet as a valid setting, attempts to set it
fail. This is exactly the issue in #42. This commit should close the loop
on future parser related issues.