Use more readable regex in manifests spec
The regex used to test the project's ability to catch syntax errors in
Puppet code was too specific for the various versions of Puppet. In the
end this contributed to extra and unneeded conditionals detracting from
code clarity among other things.

This commit uses a much more succinct regular expression that can be
trusted across major versions of Puppet.