Refactor `manifests_spec.rb` for readability
Without this commit, there is a nested if structure controlling how
deprecation warning tests will be run in relation to the version of
Puppet we're testing against. Wile effective, it could stand to be more
readable.

This commit refactora the code in question to use a "case" structure and
more explicit ranges for matching. This has improved the readability
overall while not changing the behavior of the tests themselves.

This commit also removes a useless comment that, not that we understand
how to test against Puppet 4, is no longer needed.