Refactor template logic and add tests
Without this commit, there was no validation that puppet-syntax would
terminate as expected when asked to validate an EPP template on a Puppet
version before 4.

This commit refactors the code that catches the condition and throws the
exception to be more readable. This commit also includes a new unit test
to validate and ensure this behavior.