Force a chdir back to our starting pwd after each template
In PR#47 the issue of a template executing a chdir causing all future templates
in the puppet-syntax run to fail with 'Errno::ENOENT: No such file or directory'
was rasised. We've managed to reproduce this on the command line, and fix it with
the included code, but we can only reproduce it very occasionally under rspec so
I'm submitting this as a PR without a test case.