Update gem author to Government Digital Service
I believe that gems under the `govuk` rubygems account and that are under alphagov organisation should have no individual authors but mention `Government Digital Service` as a whole.

We are a team and most of us contribute to several gems but never got to add themselves as an author. By setting this as `Government Digital Service` it's self implied that anyone that contributed is contributing for the `Government Digital Service`.