Task does non-zero exit when a syntax error or warning is found.
Prior to this fix, when the task will fail with a RunTimeError
when it encounters a failure or warning. This leads to a very ugly
Ruby stacktrace in STDERR. With this fix, the task will print the
warnings to STDOUT, and exit with a 1 exit code to indicate a
failure in the run.