Allow specifying file paths for manifests and templates too
This can be useful in various situations, like when setting up a task
for checking a specific environment only (where we might be using
e.g. the future parser, or not allowing deprecations, and so on), or
when using this as part of a CI pipeline, where you might want to check
only the changed files for syntax, to save processing time.