Raise warning if hiera check is called but disabled
Previously we got the following output from the `syntax` rake task:

```terminal
$ bundle exec rake syntax
---> syntax:manifests
---> syntax:templates
---> syntax:hiera:yaml
```

This is a bit misleading because the `syntax:hiera:yaml` task was indeed
called, but it doesn't do anything because
`PuppetSyntax.check_hiera_keys` was false. With this commit the output
changes:

```terminal
$ bundle exec rake syntax---> syntax:manifests
---> syntax:templates
---> syntax:hiera:yaml
syntax:hiera:yaml was called, but PuppetSyntax.check_hiera_keys is false. hiera syntax won't be checked
```